"revResistance" part 2
It's literally just adding that value thing here

Copied and pasted from other Pull Request

Words to read:
What does this do:
Makes engines revv slower or faster depending on how high or low the value is.
If the value is "0" or negative the engine  will default it to 10.

What does it change:
Originally a hard coded value, (10) this controls how fast the engine revvs in idle and how fast the engine accelerates.
The most notable difference is that engines in neutral rev slower, and have a bit less redline kickback, though hopefully it shouldn't be an issue. Otherwise, there shouldn't be any visible changes to current engines without a "revResistance" value
Engines with a "revResistance" value lower than 10 should notice a quick change in RPMs, rather than trying to play "catch up" and failing.

Why would this be needed:
The engine RPM is playing a game of "catch up" and sometimes the engine can't actually catch up. Most visible on cars with close gear ratios or that accelerate quickly.