[PR] Definable Engine Shutoff Winddown Rate (or whatever the hell you call it) and Definable-By-Engine Hours Factor (which cannot go below 0.25) (#935)
Allows you to define how quickly the engine slows down when not running, and not in gear. (By default it's 10, and can be set to anything... would NOT recommend negative numbers, but then again literally anything in this mod wigs out with negative numbers)
Allows you to make more rugged, or fragile and brittle engines... well, no. actually. It just changes how quickly hours get applied.