Combined areVariablesBlocking and areVariablesLocking
They are now `isVariableListTrue`, a more direct function.

If there is no list, it is treated as true. For this reason, the permanent version needs to explicitly check if there is a list because we want to use a default variable, not assume it is true.

This function is inverted from areVariablesBlocking, so where it was used is also inverted.

This function also now requires the specific list, rather than just the definition.