Fixed a bug with conditionalDefaultParts (#1401)
* Fixed a bug with conditionalDefaultParts

Fixed the bug where upon initial placement the part would not be present if the variable was defined in the rendering constants section, as opposed to the definitions constant section.

This repeats the setting the constants, but that is needed since this only executes when there is a conditional default part on a new entity.

* Rearranged defaultPart logic.

Moving the `defaultPart` after `conditionalDefaultParts` allows `defaultPart` to act as a fallback when no conditions are met. The original order resulted in the `defaultPart` overriding `conditionalDefaultParts`.