Add ProNav for vehicle targets
Add proportional navigation to guided bullets tracking vehicles. This stops bullets from "chasing" a target which has a high likelihood of missing it's target at high-aspect angles. This causes the bullet to aim where the moving target will be when the missile arrives. Forcing it to lead the target. This makes guided weapons much more effective and efficient. However it makes them much harder to evade.