Adds `random_flip` variable, based on world time
Unfortunately means this doesn't work when the `doDaylightCycle` gamerule is set to false, but I'd generally assume that shouldn't be an issue.

Useful for defining a random chance between one set of conditionalDefaultParts or the other. Examples: setting a normal defaultPart along with one CDP to switch between the two, or not having a DP defined for a 50/50 chance of getting that part or not.